// src/models/CartProduct.ts
// models/CartProduct.ts
import { Attribute } from './Attribute';

export interface CartProduct {
    id: string; // ID sản phẩm
    name: string; // Tên sản phẩm
    originalPrice: number; // Giá gốc
    discountedPrice: number; // Giá đã giảm
    quantity: number; // Số lượng sản phẩm
    attributes: Attribute[]; // Thuộc tính của sản phẩm
    
}